refactor(admin): split permission resource panel
