+++ apps/admin/src/pages/Permission/FilterForm.tsx
@@ -1,0 +1,25 @@
+import { SearchForm } from '@kagari/ui/components/ProRestful';
+import { FormItem, ProForm } from '@kagari/ui';
+import Input from '@mui/material/Input';
+import IconButton from '@mui/material/IconButton';
+import SearchIcon from '@mui/icons-material/Search';
+import React from 'react';
+
+export const FilterForm: SearchForm = function (props) {
+  const onSubmit = (formData: any) => {
+    props.handleList({ username: formData.search });
+  };
+  return (
+    <ProForm inline onSubmit={onSubmit}>
+      <FormItem
+        prop={'search'}
+        render={({ field, name }) => {
+          return <Input {...field} placeholder={name} />;
+        }}
+      />
+      <IconButton type="submit">
+        <SearchIcon color="primary" />
+      </IconButton>
+    </ProForm>
+  );
+};
