+++ apps/admin/src/router.tsx
@@ -6,7 +6,7 @@ import Auth from './layouts/Auth';
 import LoginPage from './pages/Login.page';
 import UserPage from './pages/User/User.page';
 import RolePage from './pages/Role/Role.page';
-import PermissionPage from './pages/Permission.page';
+import PermissionPage from './pages/Permission/Permission.page';
 
 export default createBrowserRouter([
   {
